dependency

+++ api/include/opentelemetry/trace/propagation/http_trace_context.h
@@ -147,7 +147,7 @@ class HttpTraceContext : public HTTPTextFormat<T> {
 
                 if (i%2==0) {
                     buf[i/2] = (uint8_t)((int)(*tid)<<4);
-                    std::cout<<(uint8_t)((int)(*tid)<<4)<<std::endl;
+                    std::cout<<((int)(*tid)<<4)<<std::endl;
                 } else {
                     buf[i/2] += (uint8_t)(*tid);
                     std::cout<<(uint8_t)(*tid)<<std::endl;
